+++ src/constants.jl
@@ -1,0 +1,160 @@
+#! format: off
+const BORDER_SOLID = (
+    tl = '┌',
+    tr = '┐',
+    bl = '└',
+    br = '┘',
+    t = '─',
+    l = '│',
+    b = '─',
+    r = '│',
+)
+const BORDER_CORNERS = (
+    tl = '┌',
+    tr = '┐',
+    bl = '└',
+    br = '┘',
+    t = ' ',
+    l = ' ',
+    b = ' ',
+    r = ' ',
+)
+const BORDER_BARPLOT = (
+    tl = '┌',
+    tr = '┐',
+    bl = '└',
+    br = '┘',
+    t = ' ',
+    l = '┤',
+    b = ' ',
+    r = ' ',
+)
+const BORDER_BOLD = (
+    tl = '┏',
+    tr = '┓',
+    bl = '┗',
+    br = '┛',
+    t = '━',
+    l = '┃',
+    b = '━',
+    r = '┃',
+)
+const BORDER_NONE = (
+    tl = ' ',
+    tr = ' ',
+    bl = ' ',
+    br = ' ',
+    t = ' ',
+    l = ' ',
+    b = ' ',
+    r = ' ',
+)
+const BORDER_BNONE = (
+    tl = Char(0x2800),
+    tr = Char(0x2800),
+    bl = Char(0x2800),
+    br = Char(0x2800),
+    t = Char(0x2800),
+    l = Char(0x2800),
+    b = Char(0x2800),
+    r = Char(0x2800),
+)
+const BORDER_DASHED = (
+    tl = '┌',
+    tr = '┐',
+    bl = '└',
+    br = '┘',
+    t = '╌',
+    l = '┊',
+    b = '╌',
+    r = '┊',
+)
+const BORDER_DOTTED = (
+    tl = '⡤',
+    tr = '⢤',
+    bl = '⠓',
+    br = '⠚',
+    t = '⠤',
+    l = '⡇',
+    b = '⠒',
+    r = '⢸',
+)
+const BORDER_ASCII = (
+    tl = '+',
+    tr = '+',
+    bl = '+',
+    br = '+',
+    t = '-',
+    l = '|',
+    b = '-',
+    r = '|',
+)
+const BORDERMAP = (
+    solid   = BORDER_SOLID,
+    corners = BORDER_CORNERS,
+    barplot = BORDER_BARPLOT,
+    bold    = BORDER_BOLD,
+    none    = BORDER_NONE,
+    bnone   = BORDER_BNONE,
+    dashed  = BORDER_DASHED,
+    dotted  = BORDER_DOTTED,
+    ascii   = BORDER_ASCII,
+)
+
+const MARKERS = (
+    circle    = '⚬',
+    rect      = '▫',
+    diamond   = '◇',
+    hexagon   = '⬡',
+    cross     = '✚',
+    xcross    = '✖',
+    utriangle = '△',
+    dtriangle = '▽',
+    rtriangle = '▷',
+    ltriangle = '◁',
+    pentagon  = '⬠',
+    star4     = '✦',
+    star5     = '★',
+    star6     = '✶',
+    star8     = '✴',
+    vline     = '|',
+    hline     = '―',
+    (+)       = '+',
+    (x)       = '⨯',
+)
+const SUPERSCRIPT = Dict(
+    # '.' => '‧',  # U+2027: Hyphenation Point
+    # '.' => '˙',  # U+02D9: Dot Above
+    # '.' => '⸳',  # U+2E33: Raised Dot
+    # '.' => '⋅',  # U+22C5: Dot Operator
+    # '.' => '·',  # U+00B7: Middle Dot
+    '.' => '⸱',  # U+2E31: Word Separator Middle Dot
+    '-' => '⁻',
+    '+' => '⁺',
+    '0' => '⁰',
+    '1' => '¹',
+    '2' => '²',
+    '3' => '³',
+    '4' => '⁴',
+    '5' => '⁵',
+    '6' => '⁶',
+    '7' => '⁷',
+    '8' => '⁸',
+    '9' => '⁹',
+)
+#! format: on
+
+const COLOR_CYCLE = (:green, :blue, :red, :magenta, :yellow, :cyan)
+
+# standard terminals seem to respect a 4:3 aspect ratio
+# unix.stackexchange.com/questions/148569/standard-terminal-font-aspect-ratio
+# retrocomputing.stackexchange.com/questions/5629/why-did-80x25-become-the-text-monitor-standard
+const ASPECT_RATIO = 4 / 3
+
+# default display size for the default BrailleCanvas (which has aspect ratio = 2) ==> (40, 15)
+const DEFAULT_HEIGHT = Ref(15)
+const DEFAULT_WIDTH = Ref(2round(Int, ASPECT_RATIO * DEFAULT_HEIGHT[]))
+
+const FSCALES = (identity = identity, ln = log, log2 = log2, log10 = log10)  # forward
+const ISCALES = (identity = identity, ln = exp, log2 = exp2, log10 = exp10)  # inverse
+const BASES = (identity = nothing, ln = "ℯ", log2 = "2", log10 = "10")
